Revert "Migrate to null safety"
This reverts commit a4198d2796c1cab9b1ef7ebf6918e11a2161470f.

+++ lib/ui/Widgets/customFilterSheet.dart
@@ -22,7 +22,7 @@ class _CustomFilterSheetState extends State<CustomFilterSheet> {
             value: index,
             groupValue: provider.transactionIndex,
             onChanged: (index) {
-              provider.setTransactionIndex(index as int);
+              provider.setTransactionIndex(index);
               Navigator.pop(context);
             },
             title: Text(
